requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
httpx==0.24.1
langchain==0.0.267
langchain-community>=0.0.1
sentence-transformers==2.2.2
pypdf==3.15.1
docx2txt==0.8
python-pptx==0.6.21
faiss-cpu==1.7.4
chromadb==0.4.6
pinecone-client==2.2.4
weaviate-client==3.24.1
tiktoken==0.5.1
pytest==7.4.0
requests==2.31.0
sse-starlette==1.6.5
